pyproject.toml
[tool.poetry]
name = "ai_core"
version = "0.1.0"
description = ""
authors = ["Tianchen Zheng <eric.tc.zheng@gmail.com>"]
readme = "README.md"
packages = [
    { include = "ai_core" }
]

[tool.poetry.dependencies]
python = "^3.11"
torch = "^2.3.0"
transformers = "^4.40.1"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.29"}
python-dotenv = "^1.0.1"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"


[tool.poetry.group.dev.dependencies]
mypy = "^1.10.0"
pytest = "^8.1.2"
ruff = "^0.4.2"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.6"
coverage = "^7.5.0"
pytest-cov = "^5.0.0"
alembic = "^1.13.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=ai_core --cov-report=term --cov-report=html --cov-report=xml"
testpaths = [
    "test",
]